Changed C casts to C++ casts in odometery_two_wheel

+++ src/odometry_two_wheel.cpp
@@ -27,19 +27,19 @@ void TwoWheelOdometry::updatePose() {
 	int32_t leftEncoderNewPos = leftEncoder.get_position();
 	int32_t rightEncoderNewPos = rightEncoder.get_position();
 	double dL =
-		(((double)leftEncoderNewPos - (double)leftEncoderLastPos) / 36000.0) *
-		6.28318;
+		static_cast<double>((((double)leftEncoderNewPos - (double)leftEncoderLastPos) / 36000.0) *
+		6.28318);
 	double dR =
-		(((double)rightEncoderNewPos - (double)rightEncoderLastPos) / 36000.0) *
-		6.28318;
-	double d = (dL + dR) / 2.0;
+		static_cast<double>((((double)rightEncoderNewPos - (double)rightEncoderLastPos) / 36000.0) *
+		6.28318);
+	double d = static_cast<double>((dL + dR) / 2.0);
 	double deltaThetaRad = (dR - dL) / (2.0 * dW);
 
-	double x = currentPose->x +
-			   (d * (double)cos(currentPose->theta + (deltaThetaRad / 2.0)));
-	double y = currentPose->y +
-			   (d * (double)sin(currentPose->theta + (deltaThetaRad / 2.0)));
-	double theta = currentPose->theta + deltaThetaRad;
+	double x = static_cast<double>(currentPose->x +
+			   (d * (double)cos(currentPose->theta + (deltaThetaRad / 2.0))));
+	double y = static_cast<double>(currentPose->y +
+			   (d * (double)sin(currentPose->theta + (deltaThetaRad / 2.0))));
+	double theta = static_cast<double>(currentPose->theta + deltaThetaRad);
 
 	currentPose->x = x;
 	currentPose->y = y;
